pyproject.toml
[project]
name = "aws-test"
version = "0.1.0"
description = ""
authors = [
    {name = "zurab",email = "z.gabrichidze099@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3 (>=1.38.3,<2.0.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "python-magic-bin (>=0.4.14,<0.5.0)"
]

[tool.poetry]
packages = [{include = "aws_test", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
